+++ IoCTools.Generator.Tests/Diagnostics/Ioc106AutoDepOpenMultiArityTests.cs
@@ -2,10 +2,10 @@ namespace IoCTools.Generator.Tests;
 
 using Xunit;
 
-public sealed class Ioc100AutoDepOpenMultiArityTests
+public sealed class Ioc106AutoDepOpenMultiArityTests
 {
     [Fact]
-    public void IOC100_fires_on_multi_arity_unbound_generic()
+    public void IOC106_fires_on_multi_arity_unbound_generic()
     {
         var source = @"
 using IoCTools.Abstractions.Annotations;
@@ -13,11 +13,11 @@ using System.Collections.Generic;
 [assembly: AutoDepOpen(typeof(IDictionary<,>))]
 ";
         var result = SourceGeneratorTestHelper.CompileWithGenerator(source);
-        result.Diagnostics.Should().Contain(d => d.Id == "IOC100");
+        result.Diagnostics.Should().Contain(d => d.Id == "IOC106");
     }
 
     [Fact]
-    public void IOC100_does_not_fire_on_single_arity_unbound_generic()
+    public void IOC106_does_not_fire_on_single_arity_unbound_generic()
     {
         var source = @"
 using IoCTools.Abstractions.Annotations;
@@ -25,11 +25,11 @@ using System.Collections.Generic;
 [assembly: AutoDepOpen(typeof(IEnumerable<>))]
 ";
         var result = SourceGeneratorTestHelper.CompileWithGenerator(source);
-        result.Diagnostics.Where(d => d.Id == "IOC100").Should().BeEmpty();
+        result.Diagnostics.Where(d => d.Id == "IOC106").Should().BeEmpty();
     }
 
     [Fact]
-    public void IOC100_message_contains_type_and_arity()
+    public void IOC106_message_contains_type_and_arity()
     {
         var source = @"
 using IoCTools.Abstractions.Annotations;
@@ -37,7 +37,7 @@ using System.Collections.Generic;
 [assembly: AutoDepOpen(typeof(IDictionary<,>))]
 ";
         var result = SourceGeneratorTestHelper.CompileWithGenerator(source);
-        var diag = result.Diagnostics.FirstOrDefault(d => d.Id == "IOC100");
+        var diag = result.Diagnostics.FirstOrDefault(d => d.Id == "IOC106");
         diag.Should().NotBeNull();
         diag!.GetMessage().Should().Contain("IDictionary").And.Contain("2");
     }